requirements.txt
# requirements.txt

streamlit
pandas
keybert
fpdf
spacy==3.5.0
plotly
sentence-transformers
wordcloud
PyPDF2
python-docx
bertopic
scikit-learn>=1.3.0  # BERTopic requires a recent version
hdbscan
networkx
pyvis
scipy
transformers
torch